Initial version extracted as a separate project
This code was extracted from a personal project with a view to one day
publishing it as a separate module. It has sat untouched for years, but
perhaps now is the time to do that.

It was originally inspired by some of the XPath matching in
Spring MVC Test
https://github.com/spring-projects/spring-framework/tree/master/spring-test
but I needed something more general for XML (and JSON), and I wanted
something that would eventually allow me to capture required or optional
values from the node too.