build: begin extracting hard-coded projectRepository.url
The hard-coded URL (set to a (Windows, mapped drive) local folder is
not very portable.
As a first step, extract it to a property, so that at least there is a
chance to override it in one's environment or build command, without
having to edit the POM itself.
Long-term, some sort of CI/CD pipeline would replace this, perhaps.