feature: improve error message for unexpected node
When asserting that a node does not exist, the error  message received
if the node does exist was confusing, it just said something like,
"node X exists". This occasionally caused confusion. "node X does not
exist" is easily interpreted as, "I expected this node but it's not
there". But the message "node X exists" looks like a verification, so
not helpful - a bit of a non-sequitur.
Improve the message so that it's clearer that the node existing is
unexpected.