style: don't use @author tag in Javadoc
It doesn't add anything, particularly given they were all the same
person, and only an email address at that. It's just clutter.
Now that the code is on GitHub, a top-level license and copyright
statement would be sufficient for copyright purposes.
And Git blame sufficient (and perhaps more accurate at the line level)
for documenting responsibility.