fix: clarify expected type in failure message
numberOfValue(int) would report that the value could not be converted
to an integer, but numberOfValue(double) did not, it just said "number".
This was for historical reasons, when there was only a single
number/double version. When the integer version was added, the old
error message should have been improved.