fix: hide log warnings by picking a logger for app
The App's logging was specified with SLF4J, but no implementation was
imposed. This approach was copied from the library, where it leaves the
user of the library to use their own implementation.
But the app should choose its own impl.
So choose logback, which is what is used for the logging in the tests.
This also stops the app logging from complaining that there is no
logging configured, thus removing some noisy log messages from every
run.