feat: out-of-bounds error matches standard List
Had disabled some platform-specific tests that depended on the
platform's precise details of a List's index out of bounds error
message.

But now, attempt to make the ListOfNode behave more like a List,
including generating error messages the same as a normal (Array)List.
This should mean the tests are less reliant on the platform-specific
details.